CSS changes to keep username box the same size
Strangely, the home/page.css is affecting the entire main page after signing in and then going back.
Thus, if certain properties are defined in home/page.css but aren't in login.css, we see the affect when going back to the main page.
May need to do something with how things are being routed in order to ultimately fix this, but for now it is fixed by simply defining everything possible in login.css.

Also tried to make the page more responsive to different viewport sizes, but haven't finalized that yet.